Changed note grid to be an array of StoredNote arraylists rather than a 2d array of StoredNotes

+++ src/StoredNote.java
@@ -1,16 +1,16 @@
 // class for storing note and position data in the Level class' noteGrid
 public class StoredNote {
     // position of the note in milliseconds
-    private float posMillis;
+    private int posMillis;
     // note object
     private Note note;
     // create a StoredNote
-    public StoredNote(float pos, int col) {
+    public StoredNote(int pos, int col) {
         this.posMillis = pos;
         this.note = new Note(col);
     }
     // get the position of the StoredNote
-    public float getPos() {
+    public int getPos() {
         return posMillis;
     }
     // get the note of the StoredNote
